Struct rusoto_kinesis_video_archived_media::KinesisVideoArchivedMediaClient

pub struct KinesisVideoArchivedMediaClient { /* fields omitted */ }

A client for the Kinesis Video Archived Media API.

Methods

impl KinesisVideoArchivedMediaClient

pub fn new(region: Region) -> KinesisVideoArchivedMediaClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> KinesisVideoArchivedMediaClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,