Crate rusoto_kinesis_video_archived_media

If you're using the service, you're probably looking for KinesisVideoArchivedMediaClient and KinesisVideoArchivedMedia.

Structs

Fragment
Represents a segment of video or other time-delimited data.

FragmentSelector
Describes the time stamp range and time stamp origin of a range of fragments.

GetHLSStreamingSessionURLInput

GetHLSStreamingSessionURLOutput

GetMediaForFragmentListInput

GetMediaForFragmentListOutput

HLSFragmentSelector
Contains the range of time stamps for the requested media, and the source of the time stamps.

HLSTimestampRange

The start and end of the time stamp range for the requested media.

This value should not be present if PlaybackType is LIVE.

The values in the HLSTimestampRange are inclusive. Fragments that begin before the start time but continue past it, or fragments that begin before the end time but continue past it, are included in the session.

KinesisVideoArchivedMediaClient
A client for the Kinesis Video Archived Media API.

ListFragmentsInput

ListFragmentsOutput

TimestampRange
The range of time stamps for which to return fragments.

Enums

GetHLSStreamingSessionURLError
Errors returned by GetHLSStreamingSessionURL

GetMediaForFragmentListError
Errors returned by GetMediaForFragmentList

ListFragmentsError
Errors returned by ListFragments

Traits

KinesisVideoArchivedMedia
Trait representing the capabilities of the Kinesis Video Archived Media API. Kinesis Video Archived Media clients implement this trait.