Crate rusoto_kinesis_video_archived_media

If you're using the service, you're probably looking for KinesisVideoArchivedMediaClient and KinesisVideoArchivedMedia.

Structs

Fragment
Represents a segment of video or other time-delimited data.

FragmentSelector
Describes the timestamp range and timestamp origin of a range of fragments.

Only fragments with a start timestamp greater than or equal to the given start time and less than or equal to the end time are returned. For example, if a stream contains fragments with the following start timestamps:

00:00:00

00:00:02

00:00:04

00:00:06

A fragment selector range with a start time of 00:00:01 and end time of 00:00:04 would return the fragments with start times of 00:00:02 and 00:00:04.

GetHLSStreamingSessionURLInput

GetHLSStreamingSessionURLOutput

GetMediaForFragmentListInput

GetMediaForFragmentListOutput

HLSFragmentSelector
Contains the range of timestamps for the requested media, and the source of the timestamps.

HLSTimestampRange

The start and end of the timestamp range for the requested media.

This value should not be present if PlaybackType is LIVE.

The values in the HLSTimestampRange are inclusive. Fragments that begin before the start time but continue past it, or fragments that begin before the end time but continue past it, are included in the session.

KinesisVideoArchivedMediaClient
A client for the Kinesis Video Archived Media API.

ListFragmentsInput

ListFragmentsOutput

TimestampRange
The range of timestamps for which to return fragments.

Enums

GetHLSStreamingSessionURLError
Errors returned by GetHLSStreamingSessionURL

GetMediaForFragmentListError
Errors returned by GetMediaForFragmentList

ListFragmentsError
Errors returned by ListFragments

Traits

KinesisVideoArchivedMedia
Trait representing the capabilities of the Kinesis Video Archived Media API. Kinesis Video Archived Media clients implement this trait.